Significant changes to index.ejs
Added a small form to add courses. Semester sections are populated with course codes and course titles in the form of an unordered list. The sections show the data corresponding to them. The form organized in a way that I can withdraw everything on the page when sending a post request.